+++ shramko.maksim/T2/DataStruct.cpp
