+++ shramko.maksim/T2/DataStruct.hpp
